Use context instead of an import
