+++ src/pages/EDCT.tsx
@@ -25,7 +25,7 @@ import AlertSnackbar, {
 import VatsimEDCTFlightPlans from "../components/EDCTFlightPlans";
 import VatsimEDCTFlightPlansViewOnly from "../components/EDCTFlightPlansViewOnly";
 import useAppContext from "../context/AppContext";
-import { useSocketContext } from "../context/SocketContext";
+import useSocket from "../hooks/useSocket";
 
 const logger = debug("edct:EDCTPage");
 
@@ -36,7 +36,7 @@ const Edct = () => {
   const [currentTime, setCurrentTime] = useState<DateTime>(DateTime.utc());
   const [snackbar, setSnackbar] = useState<AlertSnackbarProps>(null);
   const [isConnected, setIsConnected] = useState<boolean | null>(null);
-  const socket = useSocketContext();
+  const socket = useSocket();
 
   useEffect(() => {
     // Update current time every minute
